+++ src/app/guilds/guilds.component.ts
@@ -34,7 +34,7 @@ export class GuildsComponent implements OnInit, AfterViewInit {
   async ngOnInit() {
     await this.service.init();
 
-    (this.tag) ? this.loadTagLayout() : this.loadGuilds();
+    (this.tag) ? this.setTagLayout(this.tag) : this.loadGuilds();
 
     this.initialized = true;    
   }
@@ -54,8 +54,6 @@ export class GuildsComponent implements OnInit, AfterViewInit {
 
     this.guilds = guilds;
     this.savedGuilds = saved;
-
-    this.setTagLayout(this.tag);
   }
 
   ngAfterViewInit() {
@@ -114,6 +112,7 @@ export class GuildsComponent implements OnInit, AfterViewInit {
     this.tag = tag;
     
     this.resetPaginator();
+    this.loadTagLayout();
   }
 
   previousPage() {