+++ src/assets/docs/changelog.md
@@ -19,6 +19,7 @@
 **Change/Fix**:
 - `/` -> higher quality icons
 - `/dashboard/servers/:id/edit` -> users with `MANAGE_GUILD` can now edit their server listing
+- `/` -> bumped guilds only show if they have been bumped
 
 **Remove**:
 - `/api/v1/guilds` -> unnecessary guild properties