+++ src/assets/docs/changelog.md
@@ -13,13 +13,18 @@
 ### v0.0.2a
 `Upcoming`
 
+![v0.0.2a](/assets/docs/img/v0.0.2a.png)
+
 **Add**:
-- `/guilds/:id` -> report captcha 
+- `/guilds/:id` -> report captcha
+- `/api/v1/guilds` -> `guild.managerIds: string[]` -> IDs of users who can manage a server listing 
 
 **Change/Fix**:
 - `/` -> higher quality icons
 - `/dashboard/servers/:id/edit` -> users with `MANAGE_GUILD` can now edit their server listing
 - `/` -> bumped guilds only show if they have been bumped
+- `/search` -> you can now search by server ID or server name 
+- `/search` -> tags searching now filters correctly
 
 **Remove**:
 - `/api/v1/guilds` -> unnecessary guild properties
@@ -28,7 +33,7 @@
 
 ### v0.0.1a
 `18/08/2020`
-![v0.0.1a](assets/docs/img/v0.0.1a.png)
+![v0.0.1a](/assets/docs/img/v0.0.1a.png)
 
 [DList](/) and the [DList Discord Server](/server) is released!
 The first version was made across [4 live streams](https://youtube.com/ADAMJR).